Re-add navigation breadcrumbs (refs: #12134).
On #11911 it was said "They will be back in January once the donation campaign
is over", but apparently this was forgotten.

Partially reverts e54b9f7e7570d199a0729ce93237c400116b614d, that did
two (related) things at a time: I've left the responsive header height part in
case it's still useful.

I've find it too hard to guess if this is going to break the test suite,
so let's see.