Bump the Vagrant base box.
There has been a lot of changes in the base box generation code and a
lot of other changes in wip/11972-use-vagrant-in-jenkins so let's make
sure everything is in sync again.

I've uploaded the base box and provided the checksum *this* time just
in case we'll have to revert back from building the base box as part
of the build process on Jenkins' isobuilders. We'll see.