Enable the feature-11712-thunderbird APT overlay.
This will install thunderbird 1:45.8.0-3+tails2 built for amd64 on
Debian Stretch. Naturally it is a first step for the Icedove →
Thunderbird migration as well. And this package contains an AppArmor
profile (unlike the Debian Jessie package).

Refs: #12242
Will-fix: #11712, #11973

These packages also has the patch from the following upstream ticket
applied:

    https://bugzilla.mozilla.org/show_bug.cgi?id=1281959

We enable browser.download.forbid_open_with so the "Open with..."
option is hidden in the attachment download dialog, since
Thunderbird's AppArmor profile does not allow starting applications in
Tails.

Refs: #11964