Simplify and clarify.
Note that the backslash in the removed `\_${pattern}_`-style sed
patterns just says "the next character will be the delimiter", so _
becomes the delimiter. That allows you to use another delimiter than /
in compination with d, which is nice when the pattern is a path (that
tend to include /). It's a little known sed-feature. Just FYI! :)