Partially fix browser bookmarks persistence for Tor Browser 7.0a4.
Apparently this new Firefox will get the realpath of the places.sqlite
symlink, get its parent directory and create places.sqlite-{shm,wal}
there -- if the creation of these two files fail, it's like if there
are no bookmarks, and no new ones can be added (the "Add bookmarks"
dialog won't even open). So we have to allow the creation of these
files too.

Unfortunately this is not enough. If the target file of the symlink
doesn't exist it is interpreted as a corrupt database => the symlink
is deleted and replaced with a fresh bookmarks database (inside the
Tor Browser profile, not the persistent directory). So, if you already
have persistent bookmarks from before it will work, but if enable it
for the first time it won't work. This will be fixed later, and will
be documented as a known issue in Tails 3.0~rc1.