Thunderbird: disable `set -eu` for Debian's migration helpers.
At least the `do_fix_mimetypes_rdf()` helper is not designed to run
with `set -e` in particular: if there is a mimeTypes.rdf inside the
profile directory, but it doesn't contain any associations to
"icedove" or "/usr/bin/iceweasel", then the `grep` will fail in such a
way that `set -e` makes the whole script abort. When set -e is
disabled, that failing `grep` will simply result in a variable being
empty, which is the expected outcome in that situation.

Will-fix: #12580