Include the Journal last in WhisperBack bug reports.
Most of the time, when reading such a bug report, after getting some context out
of the general hardware & system information, I am mainly interested in what's
near the end of the Journal. Previously I had to search for this place in the
middle of the technical details. Now I'll be able to simply scroll to the
bottom :)