Tor Browser: kill the branding@amnesia.boum.org add-on.
This old school Firefox add-on is not compatible with e10s, so we move
its functionality into our Tor Browser wrapper instead of porting to a
WebExtension -- that would be more work, and we'd still have to do
ugly stuff to work around mandatory signing.

Refs: #12569

                                     _ " _
   _ " _                            (_\|/_)
  (_\|/_)   _ " _           _ " _    (/|\)
   (/|\)   (_\|/_) " _     (_\|/_)
            (/|\)_\|/_)     (/|\)
                 (/|\)

And yay, I found a use case for this kind of Enterprise feature that's
not about enslaving corporate computers, but for freedom, love and all
that good stuff! Namaste!

       ,,,                      ,,,
      {{{}}    ,,,             {{{}}    ,,,
   ,,, ~Y~    {{{}},,,      ,,, ~Y~    {{{}},,,
  {{}}} |/,,,  ~Y~{{}}}    {{}}} |/,,,  ~Y~{{}}}
   ~Y~ \|{{}}}/\|/ ~Y~  ,,, ~Y~ \|{{}}}/\|/ ~Y~  ,,,
   \|/ \|/~Y~  \|,,,|/ {{}}}\|/ \|/~Y~  \|,,,|/ {{}}}
   \|/ \|/\|/  \{{{}}/  ~Y~ \|/ \|/\|/  \{{{}}/  ~Y~
   \|/\\|/\|/ \\|~Y~//  \|/ \|/\\|/\|/ \\|~Y~//  \|/
   \|//\|/\|/,\\|/|/|// \|/ \|//\|/\|/,\\|/|/|// \|/
  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^