Remove a DAVE test that seems invalid to me (refs: #12005).
The page that's being used for this
test (https://tails.boum.org/install/debian/usb/) is DAVE-enabled. It doesn't
display any link to the ISO, and DAVE picks a mirror itself, so that should work
regardless of JavaScript settings. So I'm not surprised that we've been reported
several time that this test doesn't pass :)