Release process: drop instructions I think are useless.
We're now using a dedicated channel ("test") when testing, and we only use the
alpha one for actually released alphas/betas/RCs. So this line should have been
adjusted (s/alpha/test/) when I introduced the "test" channel.

But since we don't hijack the "alpha" channel anymore, I think the UDFs
on the "test" channel are harmless, and it's probably not worth removing them.

Time will tell if I missed one reason to add these lines in the first place.