Vagrant: track the latest debian-security archive for the corresponding APT sources, and not for the unrelated jessie-updates (refs: #12829).
Interestingly, this bug wasn't noticed earlier because our
tails-update-time-based-apt-snapshots-* cronjobs are set up in a way that if
everything goes fine, we're guaranteed that a snapshot of the debian archive
corresponding to the debian-security archive's latest serial always exists.

But when tails-update-time-based-apt-snapshots-* cronjobs start failing, which
is the case currently due to #12725 and/or #13134, then it may happen that this
doesn't hold true, and then the ISO build fails because we're trying to use
a snapshot that doesn't exist for the jessie-updates suite.

So, in this commit:

1. do not modify the serial used for the jessie-updates suite; this should fix
   the ISO build;
2. do update the serial used for the jessie/updates (security) suite; this
   ensures we actually install security updates in the build VM, as
   intended initially.