Rakefile: adjust enough_free_host_memory_for_ram_build? for Stretch.
The output of `free' has changed in Stretch, and we now require Stretch or newer
as the host system. It seems that we didn't notice this problem earlier as this
check is performed only when the build VM is not running yet, which rarely
happens in practice at this point.