Ensure pinentry-gtk2 run by Seahorse has the correct $DISPLAY set (refs: #12733).
As discussed in more details on https://bugs.debian.org/869416, there's a bug
somewhere in the Seahorse → gnupg-agent → pinentry-gtk2 communication, that
leads to pinentry-gtk2 being invoked with the $DISPLAY environment variable
being unset. Strangely, this does not happen for gpg → gnupg-agent →
pinentry-gtk2. Now that upstream is aware of the problem, let's apply
a temporary workaround to Tails until the root cause is fixed.