Test suite: when testing emergency shutdown, wait longer for Tails to tell us it has finished wiping the memory (refs: #13462).
The goal here is to help us understand whether #13462 is a bug in the emergency
shutdown feature or in our test suite.

In config/chroot_local-includes/usr/local/lib/initramfs-pre-shutdown-hook we
wait 2 minutes after having displayed the "Happy dumping!" message, so it should
be safe to wait 90 seconds for it to appear (and the remaining 30 seconds should
be enough for the host system to dump the VM's memory, and if that's not the
case we can bump that "sleep 120" to whatever we need).