Merge base branch earlier, i.e. in auto/config instead of auto/build (refs: #14459).
Previously, a given build from a topic branch would mix inconsistent versions
of things. Most changes done in $topic_branch..$base_branch would be taken into
account, but some would not, e.g. changes that affect:

 * everything we set up in auto/config, such as
    - copying tails-transform-mirror-url,
      gnome-shell-extension-florence-indicator
    - APT snapshots
    - debian/changelog, used by `apt-snapshots-serials prepare-build'
 * auto/build

In practice, we've been suffering from some consequences of this problem
e.g.:

 * ISO build failed on documentation branches based on master;
 * topic branches fail to build building once the APT snapshots they encode
   disappears, even though their base branch encodes newer & valid APT
   snapshots.

This commit implements a cheap and partial fix: as stated on
https://labs.riseup.net/code/issues/14459, the base branch merge still happens
too late e.g. to take into account auto/config changes done in
$topic_branch..$base_branch. Ideally we should do the base branch merge first
thing in the build process, in a manner that's isolated from other build steps,
so that *all* build code except the tiny script that performs the merge will be
in the correct state.