Fix custom polkit prompt org.boum.tails.root-terminal.policy.in and its translations (refs: #12738).
Without this, even after I've refreshed the POT and PO files, translated the new
string into French and run ./refresh-translations again, the custom polkit
prompt is not displayed even in English, which is no big surprise because the
resulting
config/chroot_local-includes/usr/share/polkit-1/actions/org.boum.tails.root-terminal.policy{,.in}
are identical i.e. no translation is integrated, and even the original string in
English is still in a <_message> block that won't be used at runtime.

This is explained by the fact that `intltool-update --desktop-style' can't
possibly work for XML files, so let's use --xml-style instead for
org.boum.tails.root-terminal.policy.

Note that I don't master awk so I don't know if the awk magics in
intltool_merge_desktop will actually *stop* after having dealt with the block
it's supposed to act on, or if that hack was assuming that we would never need
any other similar trick (which was wrong as soon as it was written, oh well) and
as a result it's going to try to handle org.boum.tails.root-terminal.policy
with --desktop-style anyway. But as we've seen already this is a no-op so
I won't bother learning awk today, and hopefully some day this code will be
rewritten in a language I understand.