Deprecate Thunderbird's preferences/0000tails.js.
We've so far only used it to set Enigmail's configuredVersion
value. We couldn't use /etc/thunderbird/pref/thunderbird.js becaue
configuredVersion must be set via a user_pref to take affect, and that
file cannot set such prefs. However, we can use the extension
override, and completely automate this.

Note: in the thunderbird wrapper we still use 0000tails.js, but in
this instance it is ok, since it's only used for a value that we'll
set each time the wrapper is run.

Will-fix: #12680