Ensure that root's bash has an appropriate PATH.
We want at least the Root Terminal to have our /usr/local/{bin,sbin}
scripts *and* overrides in its PATH. Before #12738, when the Root
Terminal was started with gksu instead of pkexec, it inherited
amnesia's PATH, but that is not the case any more.

In particular, now our /usr/local/sbin/nautilus override will be
active in the Root Terminal, which was the point of that override.

Refs: #12034, #12738