auto/config: abort if tails-custom-apt-sources failed.
Currently the build of our devel branch fails in a confusing manner.

tails-custom-apt-sources fails this way:

 + tails-custom-apt-sources
 base_branch: testing
 current_branch: devel
 In a base branch, config/base_branch must match the current branch.

… but the build goes on, and fails later when we try installing packages
that are only found in our custom APT repository, because no APT source
pointing to that repository was set up.

Let's avoid this and fail earlier when we know the build can't possibly succeed.