Re-enable the X11 testing extension: GNOME on-screen keyboard depends on it (refs: #14675).
Let's accept that our "sandbox" can be escaped via X11 and revert the changes
made in #14623: it's never been part of the threat model of this sandbox to
protect against such issues, and a11y + IBus allow basically the same kind of
breakouts anyway, so no big deal. Let's instead focus on Wayland (#12213).

This reverts commits 706e8ef83c7eb328d2e3bd633eb7aaea5c552fbb and 1f3367a450cf4b1ef8600b21d24d485810c25d7e.