Test suite: update more pictures for feature/buster.
These had to be updated by hand because we're using them in a special
way that renders --fuzzy-image-matching ineffective.

Note that I'm including *lots* of context for TorBrowserSynapticManual.png, so
that --fuzzy-image-matching does not break the tests that verify we can't see it
on the screen. The versions included in this picture have not changed since
years, so I'm not too worried. Besides, I do hope we manage to replace Synaptic
with GNOME Software at some point in Tails/Buster.