doc-impacted-by: look at paths at the new commit, not CWD.
So far we have been looking at the the documentation pages existing in
the current working directory, not at the new Git state, which is the
correct approach. It will help in this situation: imagine if at the
new Git state some page exists and is impacted, but at our current
checkout that exact file path doesn't exist any more (perhaps the file
was moved recently?). This will prompt the user to investigate and
update the relationship database accordingly (e.g. update the path, or
even removing the entry if the docs indeed were removed since the
feature was removed as well).

This also made me realize a mistake: earlier I thought the git module
worked poorly unless the current working directory was the Git root,
but that simply was the use of `Dir.glob("wiki/src/...`. Our
replacement, `git ls-tree`, is similarly affected by default, but we
fix it with the `--full-tree` option.