Test suite: allow more time for 'systemctl is-system-running' to succeed.
Hypothesis: #14772 is due to a race occurring after `step "the time
has synced"`, namely that the systemd units that will start when the
time is synced (and probably the same check is used for this as in
that step) haven't necessarily finished when we run `systemctl
is-system-running`. So let's just give it some more time.

Will-fix: #14772