Test suite: fix our findAny() vs findfailed_hook.
For findAny() it is expected that some images won't be found, so we
shouldn't use our findfailed_hook, which is about dealing with the
situation where images need to be updated.

The changes in this commit might seem unrelated, but since we need a
find_no_override() we might as well prepare them for all methods we
override, and then it also makes sense to just pass the original
method to findfailed_hook() instead of doing the ugly proxy._invoke()
stuff yet again.