Only display the survey link 1 out of n times (#14833)
The idea here is to:

- Not bother people to much with this link while still gathering enough
  data.
- Slightly prevent trolls to answer the survey in a loop.

I set n = 20 as a start which should lead to 1000 display per day
(number of boots/n). We can adjust that later on on the go depending on
the conversion rate we observe on the survey platform.

NB: People without JavaScript won't see it. But everybody has JavaScript
when landing on this page for the first time :)