Move doc to a more suitable place (refs: #14609).
Let's be consistent with how we organized the doc for automated builds:
this piece of doc is irrelevant for everyone but sysadmins + Foundation
Team, so let's move it out of the release process section.

Besides, this is not specific to lizard: any deployement of our Jenkins CI setup
is affected (even though another jenkins-jobs repo might be used, the underlying
logic is the same).