Bring in browser detection code
This is an heavy adaptation of Giorgio's code but it adapts the display
logic to the new page and also makes it more explicit.

Note that I'm taking for granted here that there will be a Chrome
version at the same time as a Firefox version which is not sure yet.