Adjust /install/download to the new download page
It won't have a "Skip download" and "Skip verification" because people
arrived on this page after clicking "Download only".

So "3. Continue installing" will be empty at first. But I still made it
point to the full installation scenario once the verification is done.
That was easier than removing "3. " and might be useful for some people
who might have landed here partly by mistake.