Vagrant: stop feeding ELF binary code to the Perl interpreter.
So in a recent apt-cacher-ng (probably the one we upgraded to when
migrating to Stretch) expire-caller.pl was turned into a symlink to
acngtool, which is pretty stupid given the extension (which makes one
think it *should* be fed to Perl), and since it is not a drop-in
replacement (something like `acngtool shrink -f` is needed).

With acngtool we're able to specify how much to clean the cache disk,
so let's try to something intelligent there.