auto/build: normalize file timestamps in wiki/src before building (refs: #14933)
The copy of the website included in the ISO image has "Posted" timestamps that
apparently match when we cloned the Git repository. This is expected since we
build the website with the rcs setting disabled. Normalizing timestamps before
building should fix this.

What's surprising though is that only a few pages seem to be affected, while
I would expect that all inlined pages without an explicit meta date should
trigger the exact same problem. But anyway, normalizing mtime should fix it
everywhere, so if it works, let's not bother investigating this further.

One potential drawback of this commit is that it may make it harder to refresh
an older, cached build of the website when we reintroduce this option in the
future:  ikiwiki will have a hard time knowing which pages it should refresh.
We'll see.