Move /install/download/openpgp into the new download page
Previously, the instructions for OpenPGP verification were in a separate
page because they implied a direct download and we didn't want to
advertise direct download for people who were not knowledgeable about
OpenPGP.

Now that direct download is advertised for everybody, we can move the
instructions here to make the life of our OpenPGP users easier.

It's at the bottom of the page and shouldn't hurt other users.

In this commit I'm also rescuing the translations we had from that page
but both the content and the translations will need some cleanup in the
next commits.