Adapt expert install to new download page
We used to point directly to the Web of Trust instructions but they are
now in a toggle. It's probably fine to point instead to the beginning of
the OpenPGP instructions because the toggle makes them very short and
the explanation of the basic verification is probably good for people to
read in order to understand its limitations.