Ask people without JavaScript to allow it temporarily
That's a shame but it's needed since we embed the display logic in
JavaScript on the page. Maybe we should have done it the other way
around but it's too late now to change our minds for a first release.

Also note that running the verification in Tor Browser with the security
slider set to high takes much much longer. Running with No Script
outside of Tor Browser is roughly the same speed as without No Script.