Test suite: fix the "I do not see ..." step's case.
The recently introduced always-on Sikuli image candidate search makes
wait() not work so great when expected to fail (as in the "do not"
case of this step) so let's use waitVanish() instead.

And, while we're at it, let's not use the FindFailed override for
waitVanish() -- it was purely a mistake that it was overridden!

Will-fix: #14929