Drop (broken) Thunderbird dedicated SocksPort (refs: #12460)
We use a dedicated SocksPort for the MUA as a trade-off: it gives poorer circuit
isolation than if we used the default SocksPort, but we were ready to compromise
on this in order to make POP-before-SMTP work. We've released Tails 3.0 with
this change 4 months ago and I've not heard about anyone being harmed by the
lack of POP-before-SMTP support, so let's give up on this one and get stronger
stream isolation in exchange.