Display a hopefully useful message with plymouth when the GDM service fails (refs: #14521)
This sets up the foundations for displaying a useful error message, but it does
not cover the case when GDM repeatedly tries to start Xorg and fails, which is
the most interesting one wrt. #14521: when this happens the gdm3 process is
still running and the service is in "active (running)" state, even once GDM
manages to put itself in a bad enough shape that it can't even retry starting
Xorg anymore.

One way to trigger such a failure is:

   systemctl kill --signal=9 gdm