Vagrant: never build the wiki early.
This has caused several issues throughout the years, the lastest
instance being the reopening of #14933, so let's just remove this
feature. The optimization this allows for when rebuilding the same
branch frequently (together with the keeprunning option) was probably
only used by me, and rarely at that, so I'd argue it is not worth the
maintenance cost.

Will-fix: #14933