Install additional software after desktop is ready
Additional Software Packages installation is currently done in GDM's PostLogin,
which locks opening of the desktop. To provide better user experience, this
commit starts the installation after the desktop startup.

This is done by a systemd user unit, triggered by desktop.target that starts a
system unit of the same name, which itself calls Additional Software Packages
installation.

To ensure that the upgrade started by a NetworkManager hook is not racy with
the installation, the hook triggers a systemd system unit which is ordered
after the installation, and waits for its state file to appear.

To have all this actually stops locking opening of the desktop, it needs a
tails-greeter version that includes commit ffc6971.

Will-fix: #9059