Drop /lib/live/mount/medium/* from the readahead list.
Commit 659442fb5f0f5aa755ce2adac6cc70c9fe263bb5 ensured such items won't be
added to that list in the future, but I forgot to actually remove them from our
current list. Let's have nightly builds and potential emergency releases benefit
from the fix as well. And if the RM for the next release forgets to update the
SquashFS sort list, well, we're covered :)

I've noticed this when a user who I had asked to test a nightly built ISO
reported he was still affected by #14964.