auto/config: use syntax compatible with "set -e" (refs: #15107).
This script has "set -e" already. I think that's currently ignored due to the
way live-build runs it, but let's not create confusion by writing code in there
that will break whenever "set -e" becomes effective. Besides, we try to use
"set -e" everywhere so let's write code that's compatible with it, which can be
useful e.g. if we copy'n'paste code around.