Disable the usr.bin.man AppArmor profile.
On Buster it breaks apparmor.service due to "profile has merged rule with
conflicting x modifiers" that's most likely caused by the "/** mrixwlk" rule vs.
our tweaks for aufs support. Usually we would add and maintain an ugly patch to
workaround this problem, but in this case I don't think it's worth our time.

As that profile says:

  # Allow basically anything in terms of file system access, subject to DAC.
  # The purpose of this profile isn't to confine man itself (that might be
  # nice in the future, but is tricky since it's quite configurable), but to
  # confine the processes it calls that parse untrusted data.

I don't think that making it safe to run man(1) on untrusted data is part of our
core mission so I'll skip that.