Test suite: improve step definition name.
I need to use this step to check for other messages that don't start with "the
destination device". "^I am told that (.*)" would feel too broad given all steps
of our test suite share a single namespace, so I'm making this one as generic as
I can in the Tails Installer context.