Tor Browser: workaround broken `7z u`?
For reasons that still are elusive to me, the `7z u` this commit
replaces with `7z a` seems to be broken. I inserted `md5sum omni.ja`
before and after the `7z u` and it printed the same value, despite `7z
u` saying that that the expected files were changed and updated.