Revert to xorg-xserver from Stretch (refs: #15232)
For #12219 we've tried upgrading to xorg-xserver 2:1.19.3-1 but that did not fix
the bug. Since then we've stuck to that version, which has a greater version
that the one in Stretch, but 1. does not get any security updates; 2. does not
track new versions from testing/sid either.

So let's get back to a saner situation and instead track the version in Stretch.