Avoid `[ -z "${@}" ]` pitfall.
The special expansion rules for $@ when it is quoted makes it
dangerous to use it with operators that expect a fixed amount of
arguments -- e.g. if it expands to `[ -z "foo" "bar" ]` we get a
syntax error at run time. So let's use $* which when quoted always
expands to a single argument.