Prefer full if statements over &&-short circuiting with set -e.
During the review it was pointed out that under `set -e`, the line:

    [ -z "${*}" ] && return

seemingly should result in an abort if the left operand fails, i.e. $*
is empty. However, the left operand of && (and ||) is considered
"explicitly tested", just like the conditional of an if-statement, so
it won't abort. Clearly this isn't common knowledge (disclaimer: I
doubted myself and had to look it up) so let's just avoid it.