Install Intel processor microcode firmware from stretch-backports (refs: #15173).
The maintainer of intel-microcode in Debian carefully uploads to
stretch-backports updates he thinks are safe for stable users. For example,
right now stretch-backports has 3.20171117.1~bpo9+1 which is the latest
available version that's not affected by the many regressions introduced by
3.20180108.1.

This commit does *not* currently give us IBRS/IBPB/STIPB microcode support for
Spectre variant 2 mitigation: the currently available firmware with that support
is too buggy. Instead, it:

 - updates microcode firmware to the latest good enough version, which usually
   brings important bugfixes;
 - paves the way for us to get this mitigation whenever it is ready in a form
   that the maintainer of intel-microcode in Debian thinks can be safely pushed
   to Debian stable users.