Test suite: use Dogtail instead of image.
This image was invalidated during our Donation campaign was running,
and then by a Tor Browser bug (#14962). Bumping images is boring so
let's rely on Dogtail instead.

To make this scenario more robust we also switch to something else
than the start page, and restart the browser between save/open in
order to prevent Dogtail e.g. finding the old page that we just saved
instead of the saved version we just opened.

Refs: #14935, #14962