Remove unhelpful paragraph
This paragraph was basically saying the same as the introduction of WoT
(and a second time in the body of the WoT section).

Keeping in mind that this information is display to everybody
downloading Tails and that a proper WoT verification is very complicated
to perform, let's explain the difference between "basic" and "WoT" only
once.

The detailed explanation of the threat at the beginning of the WoT
toggle is still and complementary with the summary of the WoT section.