Show displays configuration in TOC (#14965)
Right now we have only one subsection and it's at the very end of the
page, so I think it's fine to put it on this page and list it in TOC.

If we end up documenting more such tricks we might need a dedicated
place.

NB: We already document something similar for KeePassX but here I went
into a bit more details.