Thunderbird: explicitly disable Enigmail's Memory Hole feature.
It is already disabled in Enigmail by default, so at the moment this
is a NOP. But let's explicitly disable it (also in TorBirdy, which
controls the pref) so it is not enabled in case the default is
changed. We want to wait with enabling this until the time is right
since the current suboptimal implementations and lack of support
degrades UX too much.

Will-fix: #15201