Drop invalid "Benefit" from our Strategic Planning.
Note that the dropped claim is that "Don't have to trust your
hardware" is a *current* benefit of Tails.

Those of us attending the March 2018 monthly meeting failed to see how
Tails differs substantially from other OSes in this respect, and
suspect it was added in error, formulated the wrong way, or put in the
wrong category (perhaps it was intended as a goal?). We decided to
drop it (some might want to keep it as a goal, but we almost had that
discussion, and it seemed the way we were headed would be to greatly
have to reduce the scope of that promise).