Fix URL of mirror being mangled on failure (#15443)
replaceUrlPrefixWithRandomMirror() is not meant to be called twice on
the same elements and this leads to the URL of the mirror being mangled.

This commit introduces another class (.use-mirror-pool-on-retry) on
which replaceUrlPrefixWithRandomMirror() is called when the verification
fails. The download button with .use-mirror-pool gets hidden and
replaced by one with .use-mirror-pool-on-retry.

This not very elegant but probably the simplest solution.