Drop copying of virtualbox-guest-dkms udev rules (refs: #15532).
Starting with the virtualbox 5.2.8-dfsg-7 upload, dkms guest udev rules
have been relocated from the virtualbox-guest-dkms package to the
virtualbox-guest-utils one.

We used to install the former package in this hook to build modules and
to keep a copy of its rules file, but there's no need to do so anymore
since virtualbox-guest-utils ships it now, and we pull it through
chroot_local-packageslists anyway.