Install a virtualbox-guest-dkms-dummy package to satisfy dependencies (refs: #15532).
The virtualbox-guest-dkms package is to be installed only during the
build to prepare kernel modules, and to be removed afterwards. Since
the virtualbox 5.2.8-dfsg-7 upload, the virtualbox-guest-utils package
depends on virtualbox-guest-dkms (with a strict version) so let's
install a dummy package providing the required dependency, so that the
hook clean-up can remove the real virtualbox-guest-dkms package without
causing the virtualbox-guest-utils to go away in the process.

To be extra precise, virtualbox-guest-utils depends on:
  virtualbox-guest-dkms (= ${source:Version})
  | virtualbox-guest-source (= ${source:Version})
  | virtualbox-guest-modules

Since we're building modules from the virtualbox-guest-dkms package,
let's have a Provides field which reflects that.