Install uBlock Origin from our custom APT repo (refs: #15616)
Since #14993 we've been installing this package from sid,
but the version that's now in sid breaks the build of our devel branch.

We'll have to migrate to that new version and adjust our stuff accordingly as
part of #15023 but for now let's stick to a version that works for us.