Disable localized Atom and RSS feeds for news/.
They've been broken half of the time forever so I can't imagine anyone relying
on them. We spend way too much time trying to fix them as part of each release
process, just to see them broken again next time someone pushes a translation or
a new blog post.

Moreover, feed clients pick them up as soon as a new blog post is published and
at that time it's not been translated yet so in practice, in the huge majority
of the cases, readers will get the English version anyway.