Drop Icedove → Thunderbird migration code.
We've done the switch a year ago, in Tails 3.0, and #15693 shows
that this migration code causes issues, so let's get rid of it.

Besides, I have a hunch that this code is the only possible rationale
explanation for the dreaded "empty persistence.conf file" problem, putting aside
user intervention: it's the only piece of code we have, apart of the Tails
Persistence Setup wizard that affected users tell they did not use, that
modifies persistence.conf (refs: #10976).