Disable the uBlock logger sidebar.
As explained on https://github.com/gorhill/uBlock/releases/tag/1.16.6,
Firefox opens the sidebar by default whenever a webext that provides
one is installed, which is the case on first Tor Browser launch.

So let's do like upstream does in their released versions (the Debian build is
seen as a development one) and disable the logger sidebar.