Test suite: give 3 virtual CPUs to the system under test (refs: #15023).
… otherwise the "Watching a WebM video" test fails ("No video with supported
format and MIME type found.") and I see this in the Journal:

  Jul 04 09:25:39 amnesia tor-browser.desktop[9040]: [Child 9313, MediaPlayback #3] WARNING: Decoder=7e6a6a774200 state=DECODING_METADATA Decode metadata failed, shutting down decoder: file /var/tmp/build/firefox-a0efd2fcd6e9/dom/media/MediaDecoderStateMachine.cpp, line 379
  Jul 04 09:25:39 amnesia tor-browser.desktop[9040]: [Child 9313, MediaPlayback #3] WARNING: Decoder=7e6a6a774200 Decode error: NS_ERROR_DOM_MEDIA_METADATA_ERR (0x806e0006): file /var/tmp/build/firefox-a0efd2fcd6e9/dom/media/MediaDecoderStateMachine.cpp, line 3445

We already had to bump the number of vCPUs to 2 for the exact same reason.
I suspect that increased e10s usage has something to do with it.

I can't reproduce the bug and that video plays just fine with only 2 vCPUs on
a manual testing VM of mine, so I'll call this a test suite only issue.