Test suite: add some retry_tor magics to the WebM in Tor Browser scenario.
I've tried bumping the number of vCPUs to 4, giving the VM under test more RAM,
and it's still very unreliable for me. The Journal says:

  Jul 06 08:54:49 amnesia tor-browser.desktop[8445]: [Child 8706, MediaPlayback #1] WARNING: Decoder=72da91792200 state=DECODING_METADATA Decode metadata failed, shutting down decoder: file /var/tmp/build/firefox-a0efd2fcd6e9/dom/media/MediaDecoderStateMachine.cpp, line 379
  Jul 06 08:54:49 amnesia tor-browser.desktop[8445]: [Child 8706, MediaPlayback #1] WARNING: Decoder=72da91792200 Decode error: NS_ERROR_DOM_MEDIA_METADATA_ERR (0x806e0006): file /var/tmp/build/firefox-a0efd2fcd6e9/dom/media/MediaDecoderStateMachine.cpp, line 3445

The previous step passes fine so it's not a matter of connecting to our website,
so I have some doubt this change will help for this specific problem, but well,
it's a good thing anyway.