Revert "Test suite: give 3 virtual CPUs to the system under test (refs: #15023)."
Since I made this change, I've seen the error that this vCPUs bump was supposed
to fix even with 4 vCPUs, but I've not seen it since I've added retry_tor
magics to the relevant scenario in f255a4ede6ab285287556e39e358110e513132b1.
So even though the number of vCPUs might help reduce the risk of seeing
this error, it's not sufficient and I'd rather count on the retry_tor
magics.

This reverts commit 209de345a373163f2dd62404cd655f4920678021.