Drop obsolete chroot browser profile setup step.
This operation is a no-op nowadays (and actually triggers an error in the logs)
because we don't seed the chroot'ed browser profiles from the Tor Browser one
anymore.

And in passing, removing this makes configure_chroot_browser_profile better
scoped and more predictable: all the files it creates or modifies are now in the
chroot'ed browser's profile directory, which is exactly what one may expect
given the name of this function.